+++ docs/release_notes.md
@@ -20,10 +20,9 @@
 - [OpenAI] Support for OpenAI Realtime API with the new OpenAiRealtimeClient. The following cases are currently supported:
   - `textToSpeech`
   - `speechToSpeech`
-- [Orchestration] Added `QWEN_3_7_MAX`, `QWEN_3_7_PLUS` and `ALIBABA_TEXT_EMBEDDING_4` to model list in `OrchestrationAiModel`.
-- [Orchestration] Marked as deprecated `QWEN_3_MAX` model (retired, use newer models instead, e.g. `QWEN_3_7_MAX`).
 - [OpenAI] You can now add multiple custom headers to an `OpenAiClient` at once via `.withHeaders()`.
-- [Orchestration] Added `GEMINI_3_1_PRO_PREVIEW_EA` and `MISTRAL_MEDIUM` to model list in `OrchestrationAiModel`.
+- [OpenAI] Added `GPT_56_LUNA`, `GPT_56_SOL`, and `GPT_56_TERRA` to model list in `OpenAiModel`.
+- [Orchestration] Added `QWEN_3_7_MAX`, `QWEN_3_7_PLUS`, `ALIBABA_TEXT_EMBEDDING_4`, `GEMINI_3_1_PRO_PREVIEW_EA`, `MISTRAL_MEDIUM`, `SAP_ABAP_STARCODER2_7b`, `GPT_56_LUNA`, `GPT_56_SOL`, and `GPT_56_TERRA` to model list in `OrchestrationAiModel`.
 - [Orchestration] Reasoning content is now supported for users through `OrchestrationChatResponse.getReasoningText()` and `OrchestrationChatCompletionDelta.getDeltaReasoningText()` to see the thinking processes when using the reasoning models.
 - [Orchestration] Added support for prompt caching for new Claude models. The following messages are cacheable:
   - `SystemMessage`,